5-RèglesAgro/requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.9
pyyaml==6.0.1
python-dateutil==2.8.2
